+++ CHANGELOG.md
@@ -7,12 +7,16 @@ This file contains the list of changes made to pyjoulescope_ui.
 
 ## 1.0.32
 
-2023 Oct 18 [in progress]
+2023 Oct 20 [in progress]
 
 * Added exception catch to GL string get.
 * Added Waveform widget "auto" marker statistics text position as default  #224
 * Improved Waveform widget autoranging  #228
 * Reduced required OpenGL API version from 4.4 to 3.3.
+* Improved sidebar flyout widget #225
+  * Automatically close when mouse leaves to the right.
+  * Added vertical scroll bar to flyout.
+  * Converted settings widget to flyout, not pop-over.
 
 
 ## 1.0.31